move the components in to Home
